feat: :sparkles: flash message added for sucess and fail
