+++ src/blueprints/__init__.py
@@ -1,0 +1,17 @@
+from src.blueprints.additional import additional_bp
+from src.blueprints.admin import admin_bp
+from src.blueprints.auth import auth_bp
+from src.blueprints.gold_calculator import gold_calculator_bp
+from src.blueprints.health import health_bp
+from src.blueprints.homepage import homepage_bp
+from src.blueprints.silver_calculator import silver_calculator_bp
+
+__all__ = [
+    'additional_bp',
+    'admin_bp',
+    'auth_bp',
+    'gold_calculator_bp',
+    'health_bp',
+    'homepage_bp',
+    'silver_calculator_bp',
+]
