+++ src/calculators/__init__.py
@@ -1,0 +1,2 @@
+from src.calculators.gold_calculator import GoldCalculator
+from src.calculators.silver_calcualtor import SilverCalculator
