+++ src/calculators/gold_calculator.py
