+++ src/templates/admin_dashboard.html
@@ -45,10 +45,10 @@
         <h3><i class="fas fa-tools"></i> Admin Tools</h3>
         <div class="button-grid">
             <a href="{{ url_for('admin.manage_users') }}" class="btn btn-large"><i class="fas fa-user-cog"></i> User Management</a>
-            <a href="{{ url_for('history.history') }}" class="btn btn-large"><i class="fas fa-history"></i>Transaction History</a>
+            <a href="{{ url_for('admin.history') }}" class="btn btn-large"><i class="fas fa-history"></i>Transaction History</a>
             <a href="{{ url_for('admin.settings') }}" class="btn btn-large"><i class="fas fa-cog"></i> System Settings</a>
             <a href="{{ url_for('admin.audit_log') }}" class="btn btn-large"><i class="fas fa-clipboard-list"></i> Audit Logs</a>
-            <a href="{{ url_for('config.update_jeweller_details') }}" class="btn btn-large"><i class="fas fa-cog"></i>Jewellers Config Settings</a>
+            <a href="{{ url_for('admin.update_jeweller_details') }}" class="btn btn-large"><i class="fas fa-cog"></i>Jewellers Config Settings</a>
         </div>
     </div>
     <div class="admin-dashboard">
